+++ docs.go
@@ -1,3 +1,3 @@
-// CircleCI go package: Allow configuring a client in order to interact with
+// Package circleci go package: Allow configuring a client in order to interact with
 // the CircleCI API
-package main
+package circleci
